+++ platforms/platform-dialogflow/src/cli/interfaces.ts
